+++ pychord/constants/qualities.py
@@ -1,7 +1,6 @@
 DEFAULT_QUALITIES = [
     # chords consist of 2 notes
     ('5', (0, 7)),
-    ('sus', (0, 7)),
     # 3 notes
     ('', (0, 4, 7)),
     ('maj', (0, 4, 7)),
@@ -12,6 +11,7 @@ DEFAULT_QUALITIES = [
     ('aug', (0, 4, 8)),
     ('sus2', (0, 2, 7)),
     ('sus4', (0, 5, 7)),
+    ('sus', (0, 5, 7)),
     # 4 notes
     ('6', (0, 4, 7, 9)),
     ('7', (0, 4, 7, 10)),