+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add an example to create a MIDI file.
 - Add `QualityManager` class to overwrite default qualities.
     - Do not import `QUALITY_DICT` from modules other than quality.
+- Fix `sus` quality.
 
 ## v0.5.1