+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Forthcoming
 
+## v0.6.0
+
 - Add an example to create a MIDI file.
 - Add `QualityManager` class to overwrite default qualities.
     - Do not import `QUALITY_DICT` from modules other than quality.